requirements.txt
flask>=2.0.0
flask-cors
gunicorn
readability-lxml
lxml_html_clean
lxml
cssselect